Fix Schedule At DatePicker showing Invalid Date and resetting on poll
The studio-ui-bundle DatePicker calls toDayJs(value) without passing the
outputFormat, so DD-MM-YYYY HH:mm strings cannot be parsed — resulting in
'Invalid Date' display after selection.

Since we cannot modify the studio-ui-bundle, this works around the bug by:
- Converting scheduledAt between backend format (DD-MM-YYYY HH:mm) and
  ISO-8601 in the transformers, since ISO strings are natively parseable
  by dayjs without a format hint
- Removing outputFormat from the DatePicker so fromDayJs() produces ISO
  strings via value.format()
- Isolating RTK Query tags to prevent mutations from triggering background
  refetches that would destroy form state
- Extracting polling/progress logic into an ExecutionStatus child component
  so poll-driven re-renders don't reset the DatePicker's unconfirmed
  selection
