+++ assets/studio/js/src/modules/data-importer/data-importer-api-slice-enhanced.ts
@@ -96,6 +96,9 @@ export const {
   useBundleDataImporterConfigGetQuery,
   useBundleDataImporterConfigSaveMutation,
   useBundleDataImporterConfigHasImportFileUploadedQuery,
+  useBundleDataImporterConfigCheckImportProgressQuery,
+  useBundleDataImporterConfigStartImportMutation,
+  useBundleDataImporterConfigCancelExecutionMutation,
   useBundleDataImporterConnectionListQuery
 } = api
